+++ server_setup/__init__.py
@@ -1,3 +1,4 @@
+from server_setup.certificate import certificate
 from server_setup.nginx_site import nginx_site
 from server_setup.observability import observability
 from server_setup.postgres_cluster import postgres_backups, postgres_logging
@@ -5,6 +6,7 @@ from server_setup.postgres_db import postgres_db
 from server_setup.system import firewall, nginx, packages, services, ssh, swap
 
 __all__ = [
+    "certificate",
     "firewall",
     "nginx",
     "nginx_site",